Adding a bunch of extracted assets from the old .swf.
Not sure exactly what to use where, and may be some duplication, but all
the images / animations should be covered.

Tried to extract texture_atlas for all the animations... if not, I
believe I have sprite sheet fallbacks for the rest.

Was done rather quickly to finish before my trial of Adobe Animate CC
2019 ran out, so hopefully I caught evertything.